# openapi.yaml
openapi: 3.0.3
info:
  title: API HealthGuard
  description: |
    Backend for API HealthGuard. Sign up, create a project with a locked API
    Base URL, discover or upload endpoints, then run zero-input health checks.
    Scoring stays rule-based; Explain is optional and needs `GROQ_API_KEY`.
  version: 0.1.0
servers:
  - url: http://localhost:4000
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /projects:
    get:
      tags: [Projects]
      summary: List owned projects
      operationId: listProjects
      responses:
        "200":
          description: Newest first
          content:
            application/json:
              schema:
                type: object
                required: [projects]
                properties:
                  projects:
                    type: array
                    items:
                      $ref: "#/components/schemas/Project"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [Projects]
      summary: Create a project
      description: Base URL is locked after create. Use a new project to monitor a different API.
      operationId: createProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectRequest"
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                type: object
                required: [project]
                properties:
                  project:
                    $ref: "#/components/schemas/Project"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /projects/{projectId}:
    parameters:
      - $ref: "#/components/parameters/projectId"
    get:
      tags: [Projects]
      summary: Get a project and its endpoints
      operationId: getProject
      responses:
        "200":
          description: Project with endpoints
          content:
            application/json:
              schema:
                type: object
                required: [project]
                properties:
                  project:
                    $ref: "#/components/schemas/ProjectWithEndpoints"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/ProjectNotFound"
    patch:
      tags: [Projects]
      summary: Rename a project
      description: Changing baseUrl always returns 409. Create a new project instead.
      operationId: updateProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectRequest"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                type: object
                required: [project]
                properties:
                  project:
                    $ref: "#/components/schemas/ProjectWithEndpoints"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/ProjectNotFound"
        "409":
          description: Base URL is locked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
              example:
                error: API Base URL is locked after the project is created. Create a new project to monitor a different API.
    delete:
      tags: [Projects]
      summary: Delete a project
      description: Removes the project and its endpoints, runs, and variables.
      operationId: deleteProject
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                required: [ok]
                properties:
                  ok:
                    type: boolean
                    enum: [true]
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/ProjectNotFound"
...
  /projects/{projectId}/endpoints/{endpointId}:
    patch:
      tags: [Endpoints]
      summary: Update endpoint chain or expected status
      operationId: updateEndpoint
      parameters:
        - $ref: "#/components/parameters/projectId"
        - $ref: "#/components/parameters/endpointId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEndpointRequest"
      responses:
        "200":
          description: Updated endpoint
          content:
            application/json:
              schema:
                type: object
                required: [endpoint]
                properties:
                  endpoint:
                    $ref: "#/components/schemas/Endpoint"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Project or endpoint not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT from signup or login. Required on every /projects route and GET /auth/me.
  parameters:
    projectId:
      name: projectId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    endpointId:
      name: endpointId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    Unauthorized:
      description: Missing, malformed, or expired Bearer token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorMessage"
          examples:
            missing:
              value:
                error: Missing or malformed Authorization header
            invalid:
              value:
                error: Invalid or expired token
    ProjectNotFound:
      description: No project with this id owned by the caller
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorMessage"
          example:
            error: Project not found
    ValidationError:
      description: Zod validation failed (error is a flatten object, not a string)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationErrorBody"
  schemas:
...
    CreateProjectRequest:
      type: object
      required: [name, baseUrl]
      properties:
        name:
          type: string
          minLength: 1
        baseUrl:
          type: string
          format: uri
          example: http://localhost:5000
    UpdateProjectRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        baseUrl:
          type: string
          format: uri
          description: Accepted by the schema but always rejected with 409.
    Project:
      type: object
      required: [id, name, baseUrl, ownerId, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        baseUrl:
          type: string
          format: uri
        ownerId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ProjectWithEndpoints:
      allOf:
        - $ref: "#/components/schemas/Project"
        - type: object
          required: [endpoints]
          properties:
            endpoints:
              type: array
              items:
                $ref: "#/components/schemas/Endpoint"
    HttpMethod:
      type: string
      enum: [GET, POST, PUT, PATCH, DELETE, HEAD, OPTIONS]
    Endpoint:
      type: object
      required:
        - id
        - projectId
        - name
        - method
        - path
        - expectedStatus
        - dependsOnId
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        name:
          type: string
        method:
          $ref: "#/components/schemas/HttpMethod"
        path:
          type: string
          description: Path relative to the project Base URL
          example: /users/{id}
        expectedStatus:
          type: integer
          default: 200
        documentedStatuses:
          type: array
          items:
            type: integer
          nullable: true
        requiresAuth:
          type: boolean
        requestSchema:
          description: JSON Schema for the request body, if known
        responseSchema:
          description: JSON Schema for the response body, if known
        headers:
          type: object
          additionalProperties:
            type: string
          nullable: true
        dependsOnId:
          type: string
          format: uuid
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UpdateEndpointRequest:
      type: object
      properties:
        dependsOnId:
          type: string
          format: uuid
          nullable: true
        expectedStatus:
          type: integer
...
    ErrorMessage:
      type: object
      required: [error]
      properties:
        error:
          type: string
    ZodFlatten:
      type: object
      properties:
        formErrors:
          type: array
          items:
            type: string
        fieldErrors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
    ValidationErrorBody:
      type: object
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/ZodFlatten"
...